packages/fastapi_cookie_auth/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-cookie-auth"
version = "0.1.0"
description = "Reusable FastAPI auth helpers for bearer access tokens and HttpOnly refresh cookies."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "passlib[argon2]",
    "python-jose[cryptography]",
    "pydantic",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["."]
